refactor: Move socket code to a separate file
