+++ .gitignore
@@ -15,6 +15,7 @@ server/node_modules
 
 # misc
 server/.env
+client/.env
 .DS_Store
 .env.local
 .env.development.local
