+++ server/index.js
@@ -1,16 +1,15 @@
 const express = require('express');
 const createError = require('http-errors');
-const socketIo = require('socket.io');
 const http = require('http');
 require('dotenv/config');
 
 const app = express();
 const server = http.Server(app);
-const io = socketIo(server);
+const io = require('socket.io')(server);
 const config = require('./config/config');
 const routes = require('./api/api.js');
 const authRoutes = require('./api/auth.js');
-const gameController = require('./api/gameController');
+const socketService = require('./api/socketService');
 
 require('mongoose').connect(config.db.url, { useNewUrlParser: true });
 
@@ -24,34 +23,7 @@ app.use((req, res, next) => {
   next(createError(404));
 });
 
-const games = {};
-
-io.on('connection', (socket) => {
-  socket.on('joinGame', (playerInfo) => {
-    const availableGame = Object.values(games);
-    const openGame = availableGame.find(game => game.freeToJoin === true);
-    if (openGame) {
-      openGame.pacmanTwo = playerInfo;
-      openGame.freeToJoin = false;
-      // eslint-disable-next-line no-param-reassign
-      socket.room = openGame.roomId;
-      socket.join(openGame.roomId);
-    } else {
-      const newGame = gameController(playerInfo);
-      games.roomId = newGame;
-      // eslint-disable-next-line no-param-reassign
-      socket.room = newGame.roomId;
-      socket.join(newGame.roomId);
-    }
-  });
-
-  socket.on('disconnect', () => {
-    delete games[socket.room];
-    socket.leave(socket.room);
-  });
-
-  socket.emit('connected');
-});
+io.on('connection', socketService);
 
 // eslint-disable-next-line no-console
 server.listen(config.port, () => console.log(`Server listening on PORT: ${config.port}`));
